Change longs in vm86.h to ints to be consistent between x86-64 and i386.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1819 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/include/Asm/vm86.h
@@ -69,22 +69,22 @@ struct vm86_regs {
 /*
  * normal regs, with special meaning for the segment descriptors..
  */
-	long ebx;
-	long ecx;
-	long edx;
-	long esi;
-	long edi;
-	long ebp;
-	long eax;
-	long __null_ds;
-	long __null_es;
-	long __null_fs;
-	long __null_gs;
-	long orig_eax;
-	long eip;
+	int ebx;
+	int ecx;
+	int edx;
+	int esi;
+	int edi;
+	int ebp;
+	int eax;
+	int __null_ds;
+	int __null_es;
+	int __null_fs;
+	int __null_gs;
+	int orig_eax;
+	int eip;
 	unsigned short cs, __csh;
-	long eflags;
-	long esp;
+	int eflags;
+	int esp;
 	unsigned short ss, __ssh;
 /*
  * these are specific to v86 mode:
@@ -96,14 +96,14 @@ struct vm86_regs {
 };
 
 struct revectored_struct {
-	unsigned long __map[8];			/* 256 bits */
+	unsigned int __map[8];			/* 256 bits */
 };
 
 struct vm86_struct {
 	struct vm86_regs regs;
-	unsigned long flags;
-	unsigned long screen_bitmap;
-	unsigned long cpu_type;
+	unsigned int flags;
+	unsigned int screen_bitmap;
+	unsigned int cpu_type;
 	struct revectored_struct int_revectored;
 	struct revectored_struct int21_revectored;
 };
@@ -114,19 +114,19 @@ struct vm86_struct {
 #define VM86_SCREEN_BITMAP	0x0001
 
 struct vm86plus_info_struct {
-	unsigned long force_return_for_pic:1;
-	unsigned long vm86dbg_active:1;       /* for debugger */
-	unsigned long vm86dbg_TFpendig:1;     /* for debugger */
-	unsigned long unused:28;
-	unsigned long is_vm86pus:1;	      /* for vm86 internal use */
+	unsigned int force_return_for_pic:1;
+	unsigned int vm86dbg_active:1;       /* for debugger */
+	unsigned int vm86dbg_TFpendig:1;     /* for debugger */
+	unsigned int unused:28;
+	unsigned int is_vm86pus:1;	      /* for vm86 internal use */
 	unsigned char vm86dbg_intxxtab[32];   /* for debugger */
 };
 
 struct vm86plus_struct {
 	struct vm86_regs regs;
-	unsigned long flags;
-	unsigned long screen_bitmap;
-	unsigned long cpu_type;
+	unsigned int flags;
+	unsigned int screen_bitmap;
+	unsigned int cpu_type;
 	struct revectored_struct int_revectored;
 	struct revectored_struct int21_revectored;
 	struct vm86plus_info_struct vm86plus;
